refactor: разбил сервис статистики на 3 подмодуля
