+++ pom.xml
@@ -2,14 +2,12 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>2.7.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-
 	<name>Explore With Me</name>
 
 	<groupId>ru.practicum</groupId>
@@ -23,8 +21,9 @@
 	</properties>
 
 	<modules>
-		<module>statistic</module>
+		<!--<module>statistic</module>-->
 		<module>server</module>
+		<module>stat</module>
 	</modules>
 
 	<build>
