+++ stat/stat-dto/src/main/java/ru/practicum/Main.java
@@ -1,0 +1,7 @@
+package ru.practicum;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
