+++ stat/stat-dto/src/main/java/ru/practicum/dto/ResponseEndpointHitDto.java
@@ -1,0 +1,14 @@
+package ru.practicum.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class ResponseEndpointHitDto {
+    private Long id;
+    private String app;
+    private String uri;
+    private String ip;
+    private String timestamp;
+}
