+++ stat/stat-dto/src/main/java/ru/practicum/dto/ViewStats.java
@@ -1,0 +1,14 @@
+package ru.practicum.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+@AllArgsConstructor
+public class ViewStats {
+    private String app;
+    private String uri;
+    private Long hits;
+}
