+++ stat/stat-server/src/main/java/ru/practicum/server/StatApplication.java
@@ -1,0 +1,13 @@
+package ru.practicum.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StatApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(StatApplication.class, args);
+    }
+
+}
