+++ stat/stat-server/src/main/java/ru/practicum/server/model/EndpointHit.java
@@ -1,0 +1,25 @@
+package ru.practicum.server.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@Entity
+@Table(name = "statistics")
+public class EndpointHit {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(nullable = false)
+    private String app;
+    @Column(nullable = false)
+    private String uri;
+    @Column(nullable = false)
+    private String ip;
+    @Column(nullable = false)
+    private LocalDateTime timestamp;
+}
