+++ stat/stat-server/src/main/resources/application.properties
