+++ stat/stat-server/src/main/resources/schema.sql
