--- statistic/src/main/java/ru/practicum/statistic/StatisticApplication.java
@@ -1,13 +1,0 @@
-package ru.practicum.statistic;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StatisticApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(StatisticApplication.class, args);
-    }
-
-}
