--- statistic/src/main/java/ru/practicum/statistic/dto/ListViewStats.java
@@ -1,14 +1,0 @@
-package ru.practicum.statistic.dto;
-
-import com.fasterxml.jackson.annotation.JsonValue;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Builder
-@Getter
-public class ListViewStats {
-    @JsonValue
-    private List<ViewStats> viewStats;
-}
