--- statistic/src/main/java/ru/practicum/statistic/dto/ViewStats.java
@@ -1,14 +1,0 @@
-package ru.practicum.statistic.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class ViewStats {
-    private String app;
-    private String uri;
-    private Long hits;
-}
