--- statistic/src/main/java/ru/practicum/statistic/model/EndpointHit.java
@@ -1,25 +1,0 @@
-package ru.practicum.statistic.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "statistics")
-public class EndpointHit {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Column(nullable = false)
-    private String app;
-    @Column(nullable = false)
-    private String uri;
-    @Column(nullable = false)
-    private String ip;
-    @Column(nullable = false)
-    private LocalDateTime timestamp;
-}
