--- statistic/src/test/java/ru/practicum/statistic/StatisticApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.practicum.statistic;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class StatisticApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
